requirements.txt
SpeechRecognition>=3.8.1
nltk>=3.8
customtkinter>=5.0.0
pillow>=9.0.0
PyAudio>=0.2.11